1. implemented SetProcessWorkingSetSize() 2. fixed Set/GetPriorityClass() not to call csrss and use NtSet/QueryInformationProcess() instead
svn path=/trunk/; revision=10921

+++ reactos/include/ntos/zwtypes.h
@@ -646,8 +646,8 @@ typedef struct _QUOTA_LIMITS
 {
 	ULONG PagedPoolLimit;
 	ULONG NonPagedPoolLimit;
-	ULONG MinimumWorkingSetSize;
-	ULONG MaximumWorkingSetSize;
+	SIZE_T MinimumWorkingSetSize;
+	SIZE_T MaximumWorkingSetSize;
 	ULONG PagefileLimit;
 	TIME TimeLimit;
 } QUOTA_LIMITS, *PQUOTA_LIMITS;
